// frontend/utils/interfaces.tsx
export interface Company {
  id: number;
  uuid: string;
  user_id: number;
  owner_assigned_date: string | null;
  source_id: number | null;
  industry_id: number | null;
  parent_company_id: number | null;
  name: string;
  email: string | null;
  cf_is_emailed: string | null;
  domain: string | null;
  street: string | null;
  city: string | null;
  state: string | null;
  postal_code: string | null;
  country_id: number | null;
  import_id: number | null;
  created_by: number | null;
  next_activity_id: number | null;
  last_contacted_at: string | null;
  next_activity_date: string | null;
  deleted_at: string | null;
  created_at: string;
  updated_at: string;
  cf_country_zone: number | null;
  cf_prodint: string | null;
  cf_partnertag: string | null;
  cf_zoho_customer_id: string | null;
  cf_gst_no: string | null;
  Unnamed: number | null;
  cf_pan: string | null;
  cf_partner_id: string | null;
  cf_tan_no: string | null;
  cf_bank_name: string | null;
  cf_bank_account_number: string | null;
  cf_ifsc_code: string | null;
  cf_current_location: string | null;
  cf_office_location: string | null;
  cf_latitude: string | null;
  cf_longitude: string | null;
  cf_last_emailed_at: string | null;
  cf_nosky_partner: string | null;
}